requirements.txt
# 数据库相关
pymysql>=1.1.0
psycopg2>=2.9.0

# 数据获取
akshare>=1.12.0
tushare>=1.4.0

# 工具库
python-dotenv>=1.0.0

